Split controller.go off of the Revision reconciler
