+++ pkg/reconciler/revision/revision.go
@@ -18,48 +18,31 @@ package revision
 
 import (
 	"context"
-	"net/http"
 	"reflect"
 	"strings"
 
+	"github.com/google/go-containerregistry/pkg/authn/k8schain"
+	cachinglisters "github.com/knative/caching/pkg/client/listers/caching/v1alpha1"
+	"github.com/knative/pkg/apis/duck"
+	duckv1alpha1 "github.com/knative/pkg/apis/duck/v1alpha1"
+	"github.com/knative/pkg/controller"
+	commonlogging "github.com/knative/pkg/logging"
+	"github.com/knative/pkg/tracker"
+	"github.com/knative/serving/pkg/apis/serving/v1alpha1"
+	"github.com/knative/serving/pkg/apis/serving/v1beta1"
+	kpalisters "github.com/knative/serving/pkg/client/listers/autoscaling/v1alpha1"
+	listers "github.com/knative/serving/pkg/client/listers/serving/v1alpha1"
+	"github.com/knative/serving/pkg/reconciler"
+	"github.com/knative/serving/pkg/reconciler/revision/config"
+	"go.uber.org/zap"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/equality"
 	apierrs "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/meta"
 	"k8s.io/apimachinery/pkg/util/sets"
-	appsv1informers "k8s.io/client-go/informers/apps/v1"
-	corev1informers "k8s.io/client-go/informers/core/v1"
 	appsv1listers "k8s.io/client-go/listers/apps/v1"
 	corev1listers "k8s.io/client-go/listers/core/v1"
 	"k8s.io/client-go/tools/cache"
-
-	"github.com/google/go-containerregistry/pkg/authn/k8schain"
-	cachinginformers "github.com/knative/caching/pkg/client/informers/externalversions/caching/v1alpha1"
-	cachinglisters "github.com/knative/caching/pkg/client/listers/caching/v1alpha1"
-	"github.com/knative/pkg/apis/duck"
-	duckv1alpha1 "github.com/knative/pkg/apis/duck/v1alpha1"
-	"github.com/knative/pkg/configmap"
-	"github.com/knative/pkg/controller"
-	commonlogging "github.com/knative/pkg/logging"
-	"github.com/knative/pkg/tracker"
-	"github.com/knative/serving/pkg/apis/serving"
-	"github.com/knative/serving/pkg/apis/serving/v1alpha1"
-	"github.com/knative/serving/pkg/apis/serving/v1beta1"
-	painformers "github.com/knative/serving/pkg/client/informers/externalversions/autoscaling/v1alpha1"
-	servinginformers "github.com/knative/serving/pkg/client/informers/externalversions/serving/v1alpha1"
-	kpalisters "github.com/knative/serving/pkg/client/listers/autoscaling/v1alpha1"
-	listers "github.com/knative/serving/pkg/client/listers/serving/v1alpha1"
-	"github.com/knative/serving/pkg/deployment"
-	"github.com/knative/serving/pkg/metrics"
-	"github.com/knative/serving/pkg/network"
-	"github.com/knative/serving/pkg/reconciler"
-	"github.com/knative/serving/pkg/reconciler/revision/config"
-
-	"go.uber.org/zap"
-)
-
-const (
-	controllerAgentName = "revision-controller"
 )
 
 type changed bool
@@ -71,12 +54,6 @@ const (
 
 type resolver interface {
 	Resolve(string, k8schain.Options, sets.String) (string, error)
-}
-
-type configStore interface {
-	ToContext(ctx context.Context) context.Context
-	WatchConfigs(w configmap.Watcher)
-	Load() *config.Config
 }
 
 // Reconciler implements controller.Reconciler for Revision resources.
@@ -101,111 +78,6 @@ type Reconciler struct {
 
 // Check that our Reconciler implements controller.Reconciler
 var _ controller.Reconciler = (*Reconciler)(nil)
-
-// NewController initializes the controller and is called by the generated code
-// Registers eventhandlers to enqueue events
-// config - client configuration for talking to the apiserver
-// si - informer factory shared across all controllers for listening to events and indexing resource properties
-// queue - message queue for handling new events.  unique to this controller.
-func NewController(
-	opt reconciler.Options,
-	revisionInformer servinginformers.RevisionInformer,
-	podAutoscalerInformer painformers.PodAutoscalerInformer,
-	imageInformer cachinginformers.ImageInformer,
-	deploymentInformer appsv1informers.DeploymentInformer,
-	serviceInformer corev1informers.ServiceInformer,
-	endpointsInformer corev1informers.EndpointsInformer,
-	configMapInformer corev1informers.ConfigMapInformer,
-	buildInformerFactory duck.InformerFactory,
-) *controller.Impl {
-	transport := http.DefaultTransport
-	if rt, err := newResolverTransport(k8sCertPath); err != nil {
-		opt.Logger.Errorf("Failed to create resolver transport: %v", err)
-	} else {
-		transport = rt
-	}
-
-	c := &Reconciler{
-		Base:                reconciler.NewBase(opt, controllerAgentName),
-		revisionLister:      revisionInformer.Lister(),
-		podAutoscalerLister: podAutoscalerInformer.Lister(),
-		imageLister:         imageInformer.Lister(),
-		deploymentLister:    deploymentInformer.Lister(),
-		serviceLister:       serviceInformer.Lister(),
-		endpointsLister:     endpointsInformer.Lister(),
-		configMapLister:     configMapInformer.Lister(),
-		resolver: &digestResolver{
-			client:    opt.KubeClientSet,
-			transport: transport,
-		},
-	}
-	impl := controller.NewImpl(c, c.Logger, "Revisions", reconciler.MustNewStatsReporter("Revisions", c.Logger))
-
-	// Set up an event handler for when the resource types of interest change
-	c.Logger.Info("Setting up event handlers")
-	revisionInformer.Informer().AddEventHandler(reconciler.Handler(impl.Enqueue))
-
-	endpointsInformer.Informer().AddEventHandler(reconciler.Handler(
-		impl.EnqueueLabelOfNamespaceScopedResource("", serving.RevisionLabelKey)))
-
-	deploymentInformer.Informer().AddEventHandler(cache.FilteringResourceEventHandler{
-		FilterFunc: controller.Filter(v1alpha1.SchemeGroupVersion.WithKind("Revision")),
-		Handler:    reconciler.Handler(impl.EnqueueControllerOf),
-	})
-
-	podAutoscalerInformer.Informer().AddEventHandler(cache.FilteringResourceEventHandler{
-		FilterFunc: controller.Filter(v1alpha1.SchemeGroupVersion.WithKind("Revision")),
-		Handler:    reconciler.Handler(impl.EnqueueControllerOf),
-	})
-
-	c.tracker = tracker.New(impl.EnqueueKey, opt.GetTrackerLease())
-
-	// We don't watch for changes to Image because we don't incorporate any of its
-	// properties into our own status and should work completely in the absence of
-	// a functioning Image controller.
-
-	configMapInformer.Informer().AddEventHandler(cache.FilteringResourceEventHandler{
-		FilterFunc: controller.Filter(v1alpha1.SchemeGroupVersion.WithKind("Revision")),
-		Handler:    reconciler.Handler(impl.EnqueueControllerOf),
-	})
-
-	c.buildInformerFactory = newDuckInformerFactory(c.tracker, buildInformerFactory)
-
-	configsToResync := []interface{}{
-		&network.Config{},
-		&metrics.ObservabilityConfig{},
-		&deployment.Config{},
-	}
-
-	resync := configmap.TypeFilter(configsToResync...)(func(string, interface{}) {
-		// Triggers syncs on all revisions when configuration
-		// changes
-		impl.GlobalResync(revisionInformer.Informer())
-	})
-
-	c.configStore = config.NewStore(c.Logger.Named("config-store"), resync)
-	c.configStore.WatchConfigs(opt.ConfigMapWatcher)
-
-	return impl
-}
-
-func KResourceTypedInformerFactory(opt reconciler.Options) duck.InformerFactory {
-	return &duck.TypedInformerFactory{
-		Client:       opt.DynamicClientSet,
-		Type:         &duckv1alpha1.KResource{},
-		ResyncPeriod: opt.ResyncPeriod,
-		StopChannel:  opt.StopChannel,
-	}
-}
-
-func newDuckInformerFactory(t tracker.Interface, delegate duck.InformerFactory) duck.InformerFactory {
-	return &duck.CachedInformerFactory{
-		Delegate: &duck.EnqueueInformerFactory{
-			Delegate:     delegate,
-			EventHandler: reconciler.Handler(t.OnChanged),
-		},
-	}
-}
 
 // Reconcile compares the actual state with the desired, and attempts to
 // converge the two. It then updates the Status block of the Revision resource
